+++ packages/edit-site/src/components/sidebar-navigation-screen-templates-browse/index.js
@@ -21,6 +21,7 @@ const config = {
 		description: __(
 			'Create new template parts, or reset any customizations made to the template parts supplied by your theme.'
 		),
+		backPath: '/patterns',
 	},
 };
 
@@ -32,6 +33,7 @@ export default function SidebarNavigationScreenTemplatesBrowse() {
 		<SidebarNavigationScreen
 			title={ config[ postType ].title }
 			description={ config[ postType ].description }
+			backPath={ config[ postType ].backPath }
 		/>
 	);
 }