+++ test/e2e/specs/editor/plugins/wp-editor-meta-box.spec.js
@@ -17,7 +17,7 @@ test.describe( 'WP Editor Meta Boxes', () => {
 	} );
 
 	test( 'Should save the changes', async ( { admin, editor, page } ) => {
-		await admin.createNewPost();
+		await admin.createNewPost( { legacyCanvas: true } );
 
 		// Add title to enable valid non-empty post save.
 		await editor.canvas.type(